refactor(brands): extract generation service
